requirements.txt
streamlit>=1.29.0
PyPDF2>=3.0.1
pdfplumber>=0.10.3
pandas>=2.2.0
python-dotenv>=1.0.0
google-generativeai>=0.3.1
sentence-transformers>=2.2.2
faiss-cpu>=1.12.0
bcrypt>=4.1.1
prisma>=0.11.0
pyjwt>=2.8.0

